musl: add support for mips
