+++ src/unix/notbsd/linux/mod.rs
@@ -92,6 +92,9 @@ s! {
     }
 
     pub struct pthread_cond_t {
+        #[cfg(target_env = "musl")]
+        __align: [*const ::c_void; 0],
+        #[cfg(not(target_env = "musl"))]
         __align: [::c_longlong; 0],
         size: [u8; __SIZEOF_PTHREAD_COND_T],
     }
